Simplify configuration deserialization
Warning 1: this commit changes the current Component hashes
Warning 2: this commit changes the configuration syntax:

- the `test` property for builds was removed
- the `add_to_path` property of components is now a list of strings,
  like the global `add_to_path`

Split Configuration into multiple files, better separating concerns.
Component constructor now takes the component definition taken from
the parsed user configuration, and forwards the relevant subset to the
Build constructor.
