+++ orchestra/actions/any_of.py
@@ -1,24 +1,29 @@
+from typing import Set, Union
+
+from .action import Action
+
+
 class AnyOfAction:
     # Used to assign a unique number so that when printing a graph
     # the node does not get aliased
     INSTANCE_COUNTER = 1
 
-    def __init__(self, actions, preferred_action):
-        self.actions = actions
-        self.preferred_action = preferred_action
+    def __init__(self, actions: Set[Union[Action, "AnyOfAction"]], preferred_action: Action):
+        self.actions: Set[Union[Action, "AnyOfAction"]] = actions
+        self.preferred_action: Union[Action, "AnyOfAction"] = preferred_action
         self.unique_number = AnyOfAction.INSTANCE_COUNTER
         AnyOfAction.INSTANCE_COUNTER += 1
 
-    def add_explicit_dependency(self, dependency):
+    def add_explicit_dependency(self, dependency: Union[Action, "AnyOfAction"]):
         for action in self.actions:
             action.add_explicit_dependency(dependency)
 
     @property
-    def dependencies(self):
+    def dependencies(self) -> Set[Union[Action, "AnyOfAction"]]:
         return {a for a in self.actions}
 
     @property
-    def dependencies_for_hash(self):
+    def dependencies_for_hash(self) -> Set[Union[Action, "AnyOfAction"]]:
         return self.dependencies
 
     def is_satisfied(self):
@@ -32,4 +37,14 @@ class AnyOfAction:
         return f"AnyOf [{self.unique_number}]"
 
     def __repr__(self):
-        return f"AnyOf [{self.unique_number}]"
+        return f"Any of {{{', '.join(a.name_for_components for a in self.actions)}}}"
+
+    def __eq__(self, other):
+        if not isinstance(other, AnyOfAction):
+            return False
+
+        return self.dependencies == other.dependencies and self.preferred_action == other.preferred_action
+
+    def __hash__(self):
+        # TODO: implement properly
+        return 0
