+++ orchestra/model/_hash.py
@@ -1,0 +1,5 @@
+import hashlib
+
+
+def hash(to_hash: str) -> str:
+    return hashlib.sha1(to_hash.encode("utf-8")).hexdigest()
